requirements.txt
parsel
Pillow
PyYAML
Requests
SQLAlchemy
saucenao_api